fix(builders): make tests more legible
