+++ clients/webclients/utils/bang/builders/content_crawl/content_crawl.ts
@@ -1,10 +1,10 @@
 // brian taylor vann
 // content_crawl
 
-import { StructureRender } from "../../../type_flyweight/structure";
-import { Vector } from "../../../type_flyweight/text_vector";
-import { ContentCrawlAction } from "../../../type_flyweight/content_crawl";
-import { copy } from "../../../text_vector/text_vector";
+import { StructureRender } from "../../type_flyweight/structure";
+import { Vector } from "../../type_flyweight/text_vector";
+import { ContentCrawlAction } from "../../type_flyweight/content_crawl";
+import { copy } from "../../text_vector/text_vector";
 
 type ContentCrawl = <A>(
   template: StructureRender<A>,
